+++ modulo3/projeto-labex/labex/src/Route/Router.js
@@ -1,10 +1,10 @@
-import AdminHome from '../pages/AdminHomePage'
-import Form from '../pages/ApplicationFormPage'
-import Create from '../pages/CreateTripPage'
-import Home from '../pages/HomePage'
-import List from '../pages/ListTripsPage'
-import Login from '../pages/LoginPage'
-import Details from '../pages/TripDetailsPage'
+import AdminHome from '../pages/Admin/AdminHomePage'
+import Form from '../pages/Application/ApplicationFormPage'
+import Create from '../pages/Create/CreateTripPage'
+import Home from '../pages/Home/HomePage'
+import List from '../pages/List/ListTripsPage'
+import Login from '../pages/Login/LoginPage'
+import Details from '../pages/Trip/TripDetailsPage'
 import { BrowserRouter, Switch, Route} from 'react-router-dom'
 import Header from '../components/Header'
