+++ modulo3/projeto-labex/labex/src/components/Header.js
@@ -6,7 +6,7 @@ import { useHistory } from 'react-router-dom';
 const Headerd = styled.div`
   background-color: #091353;
   display: flex;
-  height: 70px;
+  height: 65px;
   align-items: center;
   
   
@@ -24,13 +24,13 @@ const Logo = styled.button`
 
 const Amarelo = styled.p`
   color: #FFD523;
-  font-size: 40px;
+  font-size: 38px;
   margin-left: 15px;
 `
 
 const Branco = styled.p`
   color: white;
-  font-size: 40px;
+  font-size: 38px;
 `
