+++ modulo3/projeto-labex/labex/src/pages/Application/style.js
@@ -21,8 +21,8 @@ export const Back = styled.div`
   position: absolute;
   bottom: -550px;
   right: -550px;
-  width: 1010px;
-  height: 1010px;
+  width: 900px;
+  height: 900px;
   border-radius: 50%;
   background-color: #091353;
 `
@@ -42,7 +42,7 @@ export const Div = styled.form`
 
 export const Select = styled.select`
   width: 40%;
-  height: 45px;
+  height: 40px;
   font-size: large;
   border-radius: 10px;
   
@@ -50,14 +50,14 @@ export const Select = styled.select`
 
 export const Input = styled.input`
   width: 39.5%;
-  height: 43px;
+  height: 38px;
   font-size: large;
   border-radius: 10px;
 `
 
 export const Button = styled.button`
-  width: 35%;
-  height: 40px;
+  width: 36%;
+  height: 35px;
   margin-top: 15px;
   cursor: pointer;
   font-size: large;
@@ -71,15 +71,15 @@ export const Button = styled.button`
 export const Buttons = styled.div`
   display: flex;
   justify-content: space-around;
-  width: 30%;
+  width: 40%;
   margin: 0 auto;
   margin-top: 10px;
   align-items: center;
 `
 
 export const Button2 = styled.button`
-  width: 35%;
-  height: 40px;
+  width: 40%;
+  height: 35px;
   margin-top: 15px;
   cursor: pointer;
   font-size: large;
