+++ modulo3/projeto-labex/labex/src/pages/Create/style.js
@@ -1,0 +1,56 @@
+import styled from 'styled-components';
+
+export const Form = styled.form`
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0 0 1em #091353;
+  width: 45%;
+  margin: 0 auto;
+  background-color: #577BC1;
+  border-radius: 15px;
+`
+
+export const Input = styled.input`
+  width: 60%;
+  height: 35px;
+  margin: 0 auto;
+  border-radius: 10px;
+  margin-top: 15px;
+  font-size: medium;
+  border-style: hidden;
+`
+
+export const Select = styled.select`
+  width: 60%;
+  height: 37px;
+  margin: 0 auto;
+  border-radius: 10px;
+  margin-top: 15px;
+  font-size: medium;
+  border-style: hidden;
+`
+
+export const Button = styled.button`
+  margin: 0 auto;
+  width: 35%;
+  font-size: large;
+  margin-top: 15px;
+  margin-bottom: 15px;
+  height: 30px;
+  border-style: hidden;
+  border-radius: 10px;
+  background-color: #FFD523;
+  color: #091353;
+  cursor: pointer;
+`
+export const Button2 = styled.button`
+  width: 16%;
+  font-size: large;
+  margin-top: 15px;
+  height: 30px;
+  border-style: hidden;
+  border-radius: 10px;
+  background-color: #091353;
+  color: #FFD523;
+  cursor: pointer;
+`
