+++ modulo3/projeto-labex/labex/src/pages/Home/HomePage.js
@@ -1,54 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 import { useHistory } from 'react-router-dom';
-import { Button, ButtonGroup } from '@chakra-ui/react'
-
-const Card = styled.div`
-  /* position: relative; */
-  bottom: -80px;
-  margin: 0 auto;
-  width: 50%;
-  height: 400px;
-  border-radius: 50px;
-  background-color: #FFD523;
-  margin-top: 70px;
-`
-
-const Text = styled.div`
-  margin: 20px;
-  padding-top: 10px;
-  font-size: large;
-`
-const Titulo = styled.div`
-  padding-top: 25px;
-  font-size: large;
-`
-
-const ButtonAd = styled.button`
-  width: 35%;
-  height: 40px;
-  margin-top: 15px;
-  cursor: pointer;
-  font-size: large;
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  border-radius: 10px;
-  background-color: #091353;
-  border-style: hidden;
-  color: #FFD523;
-`
-
-const Buttons = styled.div`
-  display: flex;
-  justify-content: space-around;
-  width: 30%;
-  margin: 0 auto;
-  margin-top: 10px;
-  align-items: center;
-  
-`
-
-
-
+import {Card, Text, Titulo, ButtonAd, Buttons} from './style'
 
 function HomePage() {
   const history = useHistory()
