+++ modulo3/projeto-labex/labex/src/pages/Home/style.js
@@ -1,0 +1,52 @@
+import styled from 'styled-components';
+
+export const Card = styled.div`
+  /* position: relative; */
+  bottom: -60px;
+  margin: 0 auto;
+  width: 50%;
+  height: auto;
+  border-radius: 50px;
+  background-color: #FFD523;
+  margin-top: 70px;
+`
+
+export const Text = styled.div`
+  margin: 20px;
+  padding-top: 10px;
+  font-size: large;
+  padding-bottom: 20px;
+`
+export const Titulo = styled.div`
+  padding-top: 20px;
+  font-size: large;
+`
+
+export const ButtonAd = styled.button`
+  width: 302;
+  display: flex;
+  flex-wrap: wrap;
+  height: 40px;
+  margin-top: 15px;
+  cursor: pointer;
+  font-size: large;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  border-radius: 10px;
+  background-color: #091353;
+  border-style: hidden;
+  color: #FFD523;
+  margin: 0 10px 0 10px;
+  justify-content: center;
+  align-items: center;
+`
+
+export const Buttons = styled.div`
+  display: flex;
+  justify-content: center;
+  width: auto;
+  height: auto;
+  margin: 0 auto;
+  margin-top: 10px;
+  align-items: center;
+  flex-wrap: wrap;
+`
